move valus from work page to workdetails
